# back-end/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "E-commerce API",
    "description": "An API for users to register/login and view different products.\n                    Registered users can add products to their cart and checkout to create \n                    their order. Registered users with admin role can add, edit and delete the products."
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/brands/": {
      "get": {
        "tags": [
          "Brands"
        ],
        "description": "Get all brands",
        "responses": {
          "200": {
            "description": "Successfully fetched list of brands",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "success"
                },
                "statuscode": {
                  "type": "number",
                  "example": 200
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "result": {
                      "type": "string",
                      "example": "Brands found"
                    },
                    "brands": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Brand"
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "schema": {
              "$ref": "#components/schemas/ServerError"
            }
          }
        }
      },
      "post": {
        "tags": [
          "Brands"
        ],
        "description": "Create new brand",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "description": "Brand name to create",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "Example name"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Brand created successfully.",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "success"
                },
                "statuscode": {
                  "type": "number",
                  "example": 201
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "result": {
                      "type": "string",
                      "example": "Brand created"
                    },
                    "brand": {
                      "$ref": "#/components/schemas/Brand"
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "Bad request — missing required fields.",
            "schema": {
              "$ref": "#components/schemas/MissingParameters"
            }
          },
          "401": {
            "description": "User provided invalid token",
            "schema": {
              "$ref": "#components/schemas/FailedAuthentication"
            }
          },
          "403": {
            "description": "User is not admin",
            "schema": {
              "$ref": "#components/schemas/NotAdmin"
            }
          },
          "500": {
            "description": "Internal server error.",
            "schema": {
              "$ref": "#components/schemas/ServerError"
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/brands/{id}": {
      "delete": {
        "tags": [
          "Brands"
        ],
        "description": "Delete brand by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID of the brand to update",
            "example": 1
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Brand deleted successfully",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "success"
                },
                "statuscode": {
                  "type": "number",
                  "example": 200
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "result": {
                      "type": "string",
                      "example": "Brand deleted"
                    },
                    "brand": {
                      "$ref": "#/components/schemas/Brand"
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "Bad request — missing required fields.",
            "schema": {
              "$ref": "#components/schemas/MissingParameters"
            }
          },
          "401": {
            "description": "User provided invalid token",
            "schema": {
              "$ref": "#components/schemas/FailedAuthentication"
            }
          },
          "403": {
            "description": "User is not admin",
            "schema": {
              "$ref": "#components/schemas/NotAdmin"
            }
          },
          "404": {
            "description": "Brand not found",
            "schema": {
              "$ref": "#components/schemas/ResourceNotFound"
            }
          },
          "500": {
            "description": "Internal server error.",
            "schema": {
              "$ref": "#components/schemas/ServerError"
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "Brands"
        ],
        "description": "Update brand by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID of the brand to update",
            "example": 1
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Product fields to update",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "New name"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Brand updated successfully",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "success"
                },
                "statuscode": {
                  "type": "number",
                  "example": 200
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "result": {
                      "type": "string",
                      "example": "Brand updated"
                    },
                    "product": {
                      "$ref": "#/components/schemas/Brand"
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "Bad request — missing required fields.",
            "schema": {
              "$ref": "#components/schemas/MissingParameters"
            }
          },
          "401": {
            "description": "User provided invalid token",
            "schema": {
              "$ref": "#components/schemas/FailedAuthentication"
            }
          },
          "403": {
            "description": "User is not admin",
            "schema": {
              "$ref": "#components/schemas/NotAdmin"
            }
          },
          "404": {
            "description": "Brand not found",
            "schema": {
              "$ref": "#components/schemas/ResourceNotFound"
            }
          },
          "500": {
            "description": "Internal server error.",
            "schema": {
              "$ref": "#components/schemas/ServerError"
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Brand": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Example name"
          }
        }
      },
...
    }
  }
}